backend/requirements.txt
aioice==0.9.0
aiortc==1.9.0
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
av==12.2.0
bidict==0.23.1
billiard==3.6.4.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
crayons==0.4.0
cryptography==42.0.8
diff-match-patch==20230430
distlib==0.3.8
Django==5.1
django-cors-headers==4.4.0
django-import-export==4.0.1
django-model-utils==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
fffw==4.0.0
filelock==3.14.0
Flask==3.0.3
Flask-SocketIO==5.3.6
google-crc32c==1.5.0
h11==0.14.0
hyperlink==21.0.0
idna==2.10
ifaddr==0.2.0
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.2.4
MarkupSafe==2.1.5
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyee==11.1.0
PyJWT==2.9.0
pylibsrtp==0.10.0
pymediainfo==5.1.0
pyOpenSSL==24.1.0
pyotp==2.9.0
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
requests==2.25.1
service-identity==24.1.0
simple-websocket==1.0.0
sqlparse==0.5.0
tablib==3.5.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==1.26.18
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
virtualenv==20.26.1
wcwidth==0.2.13
Werkzeug==3.0.3
wsproto==1.2.0
zope.interface==7.0.1
